fix(jetbrains): address repo CLI review feedback
