+++ packages/kilo-jetbrains/backend/src/main/kotlin/ai/kilocode/backend/cli/KiloBackendCliManager.kt
@@ -89,8 +89,9 @@ class KiloBackendCliManager(
         forceExtract = false
         if (!KiloProps.pinned()) {
             if (force) log.info("Force re-extracting local repo CLI ${KiloProps.cliVersion()}")
+            val cli = KiloRepoCli.extract(force)
             onProgress(CliDownload(100, KiloProps.cliVersion(), KiloCliPlatform.current()))
-            return KiloRepoCli.extract(force)
+            return cli
         }
         if (force) log.info("Force re-downloading CLI ${KiloProps.cliVersion()}")
         return KiloCliDownloader(log = log).resolve(KiloProps.cliVersion(), force, onProgress)
