+++ packages/kilo-jetbrains/script/build-version.sh
@@ -86,7 +86,7 @@ if [[ ! -d "$plugin" ]]; then
   exit 1
 fi
 
-if grep -q '^kilo\.cli\.pinned=false[[:space:]]*$' "$plugin/gradle.properties"; then
+if grep -Eq '^[[:space:]]*kilo\.cli\.pinned[[:space:]]*=[[:space:]]*false[[:space:]]*$' "$plugin/gradle.properties"; then
   echo "kilo.cli.pinned=false is a dev-only mode and cannot be released. Set kilo.cli.pinned=true before building a version." >&2
   exit 1
 fi
